(chore:polygon-queue) rename snapshotQueue->polygonQueue
